[GREEN] change implementation edit team profile

+++ team_management/views.py
@@ -21,7 +21,7 @@ class TeamManagementViewSet(mixins.CreateModelMixin,
 			'logo': request.data.get('logo'),
 		}
 		return team_data
-		
+
 	def create(self, request, *args, **kwargs):
 		team_data = self.get_team_data_from_request(request)
 		team_serializer = TeamSerializers(data=team_data)
@@ -34,21 +34,22 @@ class TeamManagementViewSet(mixins.CreateModelMixin,
 		return Response(data={"error": "Please make sure your [team name] is exist!"}, status=status.HTTP_400_BAD_REQUEST)
 
 	def update(self, request, *args, **kwargs):
-		team_data = self.get_team_data_from_request(request)
-		team_serializer = TeamSerializers(data=team_data)
-		if team_serializer.is_valid():
-			team = Team.objects.get(pk=kwargs['pk'])
-			
-			# Remove old image
-			if (team.logo and os.path.exists(team.logo.path)):
-				os.remove(team.logo.path)
+		team = Team.objects.get(pk=kwargs['pk'])
+		
+		if (request.data.get('name') is not None):
+			return Response(data={"error": "Team name cannot be changed!"}, status=status.HTTP_400_BAD_REQUEST)
+		
+		# remove old image
+		if (request.data.get('logo') is not None and team.logo and os.path.exists(team.logo.path)):
+			os.remove(team.logo.path)
+			team.logo = request.data.get('logo')
 
-			team.name = team_data['name']
-			team.description = team_data['description']
-			team.logo = team_data['logo']
-			team.save()
+		if (request.data.get('description') is not None):
+			team.description = request.data.get('description')
+		
+		team.save()
 
-			serialized_obj = TeamSerializers(team)
-			return Response(data=serialized_obj.data, status=status.HTTP_200_OK)
+		serialized_obj = TeamSerializers(team)
+		return Response(data=serialized_obj.data, status=status.HTTP_200_OK)
 			
-		return Response(data={"error": "Please make sure your [team name] is exist!"}, status=status.HTTP_400_BAD_REQUEST)
+
